% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{calculate_es}
\alias{calculate_es}
\title{calculate_es}
\usage{
calculate_es(data, n.e, mean.e, sd.e, n.c, mean.c, sd.c, type, studlab)
}
\arguments{
\item{data}{.}

\item{n.e}{.}

\item{mean.e}{.}

\item{sd.e}{.}

\item{n.c}{.}

\item{mean.c}{.}

\item{sd.c}{.}

\item{type}{.}

\item{studlab}{.}
}
\value{

}
\description{
Calculate...
}
\examples{

}
