% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{combine_between}
\alias{combine_between}
\title{combine_between}
\usage{
combine_between(x, sd, n)
}
\arguments{
\item{x}{A vector of means}

\item{sd}{A vector of standard deviations}

\item{n}{A vector of number of subjects}
}
\value{

}
\description{
Combining mean and standard devation from independant subgroups
within a study. This is computing a composite score by using the summary data
from the subgroups to recreate the data for the study as a whole. This summary
data can then be later used to compute the effect size and variance.
}
\examples{

}
