% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{combine_within}
\alias{combine_within}
\title{combine_within}
\usage{
combine_within(es, var_es, r)
}
\arguments{
\item{es}{A vector of effect sizes}

\item{var_es}{A vector of effect size variances}

\item{r}{The correlation}
}
\value{

}
\description{
Combine effect sizes from the same within-subjects study. An
example is a crossover study with multiple intervention conditions.
}
\examples{

}
