% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{es_between}
\alias{es_between}
\title{es_between}
\usage{
es_between(x1, x2, sd1, sd2, n1, n2)
}
\arguments{
\item{x1}{Mean of group 1}

\item{x2}{Mean of group 2}

\item{sd1}{Standard deviation of group 1}

\item{sd2}{Standard deviation of group 2}

\item{n1}{Number of subjects in group 1}

\item{n2}{Number of subjects in group 2}
}
\value{

}
\description{
Calculate effect size (md, d, g) from a between-subjects design
 of independent groups (e.g. a parallel randomised controlled trial).
}
\examples{

}
