% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimates_vcov.R
\name{new_estimates_vcov}
\alias{new_estimates_vcov}
\title{Low-level constructor for estimates_vcov}
\usage{
new_estimates_vcov(estimates, vcov, row_map = NULL)
}
\description{
Low-level constructor for estimates_vcov
}
\keyword{internal}
