% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{pool_between}
\alias{pool_between}
\title{pool_between}
\usage{
pool_between(data, n.e, mean.e, sd.e, n.c, mean.c, sd.c, type, studlab)
}
\arguments{
\item{data}{A dataframe}

\item{n.e}{Column name: number of subjects in experimental group}

\item{mean.e}{Column name: mean for experimental group}

\item{sd.e}{Column name: Standard deviation for experimental group}

\item{n.c}{Column name: number of subjects in control group}

\item{mean.c}{Column name: mean for control group}

\item{sd.c}{Column name: Standard deviation for control group}

\item{type}{Column name: type of study.}

\item{studlab}{Column name: study label}
}
\value{

}
\description{
Takes a dataframe of means and standard deviations and returns
a dataframe with pooled means and standard deviations for between-subject
studies. This function is normally run before calculating effect sizes.
}
\examples{

}
