% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{pool_studies}
\alias{pool_studies}
\title{pool_studies}
\usage{
pool_studies(data, n.e, mean.e, sd.e, n.c, mean.c, sd.c, type, studlab)
}
\arguments{
\item{data}{.}

\item{n.e}{.}

\item{mean.e}{.}

\item{sd.e}{.}

\item{n.c}{.}

\item{mean.c}{.}

\item{sd.c}{.}

\item{type}{.}

\item{studlab}{.}
}
\value{

}
\description{
Takes a dataframe of means, standard deviations, number of
subjects in each group, and study type (either 'within' or 'between') and
returns a dataframe of study effect sizes (that are pooled across studies
where necessary). This can then be used for meta analysis.
}
\examples{

}
