% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{pool_within}
\alias{pool_within}
\title{pool_within}
\usage{
pool_within(
  data,
  md = NULL,
  variance_md = NULL,
  se_md = NULL,
  d = NULL,
  variance_d = NULL,
  se_d = NULL,
  g = NULL,
  variance_g = NULL,
  se_g = NULL,
  type,
  studlab
)
}
\arguments{
\item{data}{.}

\item{md}{.}

\item{variance_md}{.}

\item{se_md}{.}

\item{d}{.}

\item{variance_d}{.}

\item{se_d}{.}

\item{g}{.}

\item{variance_g}{.}

\item{se_g}{.}

\item{type}{.}

\item{studlab}{.}
}
\value{

}
\description{
Pool...
}
\examples{

}
