% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/velocity.R
\docType{data}
\name{velocity}
\alias{velocity}
\title{Example data for meta analysis}
\format{
A data frame with 38 rows and 9 variables:
\describe{
  \item{study}{Study author and yeat}
  \item{n.c}{Number of subjects in the control group}
  \item{mean.c}{}
  \item{sd.c}{}
  \item{n.e}{}
  \item{mean.e}{}
  \item{sd.e}{}
  \item{type}{}
  \item{description}{}
}
}
\usage{
velocity
}
\description{
A dataset containing the results for velocity training studies
}
\keyword{datasets}
