% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rma_helpers.R
\name{rma_mv_helper}
\alias{rma_mv_helper}
\title{Run rma.mv on an estimates_vcov object}
\usage{
rma_mv_helper(object, yi, V = NULL, ...)
}
\arguments{
\item{object}{An estimates_vcov object}

\item{yi}{Formula or bare column name specifying the effect sizes (e.g., \code{estimate})}

\item{V}{Variance-covariance matrix (defaults to the vcov from object)}

\item{...}{Additional arguments passed to \code{\link[metafor:rma.mv]{metafor::rma.mv()}}, such as \code{random}, \code{mods}, etc.}
}
\value{
An object of class \code{rma.mv} as returned by \code{\link[metafor:rma.mv]{metafor::rma.mv()}}
}
\description{
A convenience wrapper for \code{\link[metafor:rma.mv]{metafor::rma.mv()}} that automatically extracts
the estimates data frame and variance-covariance matrix from an estimates_vcov object.

This function passes all arguments directly to \code{\link[metafor:rma.mv]{metafor::rma.mv()}}, but
handles the \code{data} and \code{V} arguments automatically.
}
\examples{
\dontshow{if (requireNamespace("metafor", quietly = TRUE) && requireNamespace("randomizr", quietly = TRUE) && requireNamespace("estimatr", quietly = TRUE)) withAutoprint(\{ # examplesIf}
library(dplyr)
library(randomizr)
library(estimatr)

set.seed(123)
dat_1 <- data.frame(Z = complete_ra(50, num_arms = 2), Y = rnorm(50))
dat_2 <- data.frame(Z = complete_ra(100, num_arms = 3), Y = rnorm(100))
dat_3 <- data.frame(Z = complete_ra(200, num_arms = 4), Y = rnorm(200))

fit_1 <- lm_robust(Y ~ Z, data = dat_1)
fit_2 <- lm_robust(Y ~ Z, data = dat_2)
fit_3 <- lm_robust(Y ~ Z, data = dat_3)

prepped_fits <- bind_rows(
  study_1 = prep_fit(fit_1, term = "ZT2"),
  study_2 = prep_fit(fit_2, term = c("ZT2", "ZT3")),
  study_3 = prep_fit(fit_3, term = c("ZT2", "ZT3", "ZT4")),
  .id = "study"
)
ev <- as_estimates_vcov(prepped_fits)

ev |> rma_mv_helper(yi = estimate, random = ~ 1 | id)
ev |> rma_mv_helper(yi = estimate, mods = ~ study, random = ~ 1 | id)
ev |>
  filter(study != "study_1") |>
  rma_mv_helper(yi = estimate, random = ~ 1 | id)
\dontshow{\}) # examplesIf}
}
