% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rma_helpers.R
\name{rma_uni_helper}
\alias{rma_uni_helper}
\title{Run rma.uni on an estimates_vcov object}
\usage{
rma_uni_helper(object, yi, vi = NULL, ...)
}
\arguments{
\item{object}{An estimates_vcov object}

\item{yi}{Formula or numeric vector specifying the effect sizes (passed to rma.uni)}

\item{vi}{Formula or numeric vector specifying the variances (defaults to diag(vcov))}

\item{...}{Additional arguments passed to \code{\link[metafor:rma.uni]{metafor::rma.uni()}}}
}
\value{
An object of class \code{rma.uni} as returned by \code{\link[metafor:rma.uni]{metafor::rma.uni()}}
}
\description{
A convenience wrapper for \code{\link[metafor:rma.uni]{metafor::rma.uni()}} that automatically extracts
the data and variance estimates from an estimates_vcov object.

Note: This function uses the diagonal of the vcov matrix as the variance
estimates. If you have correlated estimates, use \code{\link[=rma_mv_helper]{rma_mv_helper()}} instead.
}
\examples{
\dontrun{
library(metafor)

# Simple univariate meta-analysis
ev |> rma_uni_helper(yi = estimate)

# With moderators
ev |> rma_uni_helper(yi = estimate, mods = ~ country)
}

}
